Corrected a typo in the prescription controller file; added features to retrieve the dependency data from external services; added feature to send metric data id dependencies are valid; if POST is sucessfull in metrics service, inserts a new document in mongoDB collection
